Last commit before major refactoring.
I have decided to implement this project from scratch, this time as a standalone
project that does not depend on rllib. New implementation is in
lddp_multivariate_gaussian file (a reference implementation is in
lddp_1d_gaussian_reference_implementation). I may delete the other files in
this project so I am doing one final commit before all that.